Refactor backend tag into separate trait
